# tests/storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mik API",
        "description": "\u7c73\u5ba2\u4e4b\u5bb6API",
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/sms/code": {
            "post": {
                "tags": [
                    "SMS"
                ],
                "summary": "\u53d1\u9001\u9a8c\u8bc1\u77ed\u4fe1",
                "parameters": [
                    {
                        "name": "telephone",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u53d1\u9001\u77ed\u4fe1\u8ba4\u8bc1\u7801"
                    }
                }
            }
        },
...
    },
...
}